Move packages to app directory
