+++ internal/api/handlers/create.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"net/http"
 
-	"github.com/iubondar/url-shortener/internal/storage"
+	"github.com/iubondar/url-shortener/internal/app/storage"
 )
 
 type CreateIdHandler struct {
